requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.4.0
python-dateutil==2.8.2
pytz==2024.1

